feat(github): wire debounce and cooldown retry into trigger evaluation
Events blocked by cooldown guards were silently dropped — a review
arriving while a fix is running would be lost. Debounce and cooldown
retry use the same mechanism: defer via TriggerDebouncer with
last-write-wins semantics, then re-evaluate after the delay.

- Add retryable/retry_after_seconds to GuardResult with named constants
- Add TriggerDeferredResult and optional debouncer to EvaluateWebhookHandler
- Add on_fire callback so PR comments are posted for deferred triggers
- Clean preset names (drop redundant repo suffix)
- 7 new tests covering debounce, retry, and guard retryability

+++ packages/aef-domain/src/aef_domain/contexts/github/slices/evaluate_webhook/debouncer.py
@@ -30,7 +30,7 @@ class TriggerDebouncer:
     async def debounce(
         self,
         key: str,
-        delay_seconds: int,
+        delay_seconds: float,
         callback: Callable[[], Coroutine[Any, Any, Any]],
     ) -> None:
         """Schedule callback after delay. Resets if called again with same key.
